+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 
+* 2020-10-03: v0.0.3 Node API with Forth instances
 * 2020-09-26: v0.0.2 Published to NPM
 * 2020-09-17: v0.0.1 First working version
 * 2020-08-08: Had the idea, first files and outline of classes